requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ─── Data Processing ──────────────────────────────────────────────────────────
pandas>=2.2.0

# ─── Vector Search ────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0
numpy>=1.26.0
fastembed>=0.2.7

# ─── AI / LLM ─────────────────────────────────────────────────────────────────
openai>=1.30.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.1.0

# ─── Web Interface ────────────────────────────────────────────────────────────
streamlit>=1.34.0
requests>=2.31.0

# ─── Environment & Config ─────────────────────────────────────────────────────
python-dotenv>=1.0.0
watchdog>=4.0.0
